init datastar web ui
