+++ .vscode/settings.json
@@ -1,13 +1,7 @@
 {
-    "cSpell.words": [
-        "isrc",
-        "pdoc",
-        "spotdl"
-    ],
+    "cSpell.words": ["datastar", "isrc", "pdoc", "spotdl"],
     "python.formatting.provider": "black",
-    "python.testing.pytestArgs": [
-        "tests"
-    ],
+    "python.testing.pytestArgs": ["tests"],
     "python.testing.unittestEnabled": false,
     "python.testing.pytestEnabled": true
 }
