+++ pyproject.toml
@@ -51,6 +51,7 @@ dependencies = [
     "syncedlyrics>=1.0.1,<2",
     "soundcloud-v2>=1.6.0,<2",
     "websockets~=14.1",
+    "datastar-py>=0.6.3",
 ]
 
 [project.urls]
