+++ spotdl/utils/config.py
@@ -150,7 +150,10 @@ def get_web_ui_path() -> Path:
     - If the web-ui directory does not exist, it will be created.
     """
 
-    web_ui_path = get_spotdl_path() / "web-ui"
+    # web_ui_path = get_spotdl_path() / "web-ui"
+    # web_ui_path = get_spotdl_path() / "src" / "spotdl" / "web" / "static"
+    web_ui_path = Path(__file__).parent.parent / "web" / "static"
+    print("Web UI path:", web_ui_path)
 
     if not web_ui_path.exists():
         os.mkdir(web_ui_path)
