+++ spotdl/web/components/search_input.html
@@ -1,0 +1,11 @@
+<div class="flex flex-row items-center space-x-4 w-full">
+    <input type="text" :placeholder="placeHolder" class="input input-bordered w-full text-base-content"
+        v-model="sm.searchTerm.value" @keyup.enter="lookUp(sm.searchTerm.value)" />
+    <button v-if="dm.loading.value" class="btn btn-square btn-primary loading"></button>
+    <button v-else class="btn btn-square btn-primary" @click="lookUp(sm.searchTerm.value)">
+        <!-- zero-width space - this btn class style breaks with only a svg inside -->
+
+        <Icon v-if="sm.isValidURL(sm.searchTerm.value)" icon="clarity:download-line" class="h-6 w-6" />
+        <Icon v-else icon="clarity:search-line" class="h-6 w-6" />
+    </button>
+</div>
