+++ spotdl/web/static/assets/favicon_package_v0.16/browserconfig.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<browserconfig>
+    <msapplication>
+        <tile>
+            <square150x150logo src="/mstile-150x150.png"/>
+            <TileColor>#00a300</TileColor>
+        </tile>
+    </msapplication>
+</browserconfig>
